move FileTable to explore package
