+++ lib/dataTableHelpers.tsx
@@ -1,7 +1,3 @@
-import _ from 'lodash';
-import * as React from 'react';
-
-import ExpandableText from '../components/ExpandableText';
 import { getDefaultDataTableStyle } from '../packages/data-portal-table/src/libs/helpers';
 
 export function getDataSchemaDataTableStyle() {
@@ -25,16 +21,4 @@ export function getDataSchemaDataTableStyle() {
             },
         },
     };
-}
-
-export function truncatedTableCell<T>(cellData: T) {
-    //@ts-ignore
-    const selector = this.selector;
-    const value = _.isFunction(selector)
-        ? selector(cellData)
-        : cellData[selector as keyof T];
-
-    return value ? (
-        <ExpandableText fullText={value} truncateProps={{ lines: 4 }} />
-    ) : null;
 }
