+++ packages/data-portal-commons/src/components/ViewDetailsModal.tsx
@@ -1,12 +1,12 @@
 import _ from 'lodash';
 import React from 'react';
 import { Button, Modal } from 'react-bootstrap';
-import styles from './common.module.scss';
 import Tooltip from 'rc-tooltip';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPlusCircle } from '@fortawesome/free-solid-svg-icons';
 
-import { IEnhancedDataTableColumn } from '../packages/data-portal-table/src/components/EnhancedDataTable';
+import { IEnhancedDataTableColumn } from '../../../data-portal-table/src/components/EnhancedDataTable';
+import styles from './ViewDetailsModal.module.scss';
 
 interface IViewDetailsModalProps<CellData> {
     cellData?: CellData;
