+++ packages/data-portal-explore/src/assets/img_channel_map_r4.json
