+++ packages/data-portal-explore/src/assets/isbcgc-mappings.json
