+++ packages/data-portal-explore/src/assets/minerva-story-mappings.json
