+++ packages/data-portal-explore/src/components/common.module.scss
@@ -1,3 +1,4 @@
+/* TODO duplicate style */
 .clickable {
     cursor: pointer;
     color: #007bff;
