+++ packages/data-portal-utils/src/libs/file.ts
@@ -1,0 +1,36 @@
+import Path from 'path';
+
+export function getFileBase(filename: string) {
+    return Path.basename(filename);
+}
+
+export function getFileExtension(filename: string) {
+    return Path.extname(filename);
+}
+
+export function getFilenameWithoutExtension(base: string) {
+    return base.includes('.') ? base.slice(0, base.lastIndexOf('.')) : base;
+}
+
+export function truncateFilename(
+    filename: string,
+    leadThreshold: number = 10,
+    trailThreshold: number = 5
+) {
+    const base = getFileBase(filename);
+    const ext = getFileExtension(filename);
+    const name = getFilenameWithoutExtension(base);
+
+    let displayValue = base;
+
+    if (name.length > leadThreshold + trailThreshold) {
+        // get the first <leadThreshold> characters of the name
+        const lead = name.slice(0, leadThreshold);
+        // get the last <trailThreshold> characters of the name
+        const trail = name.slice(-trailThreshold);
+        // always keep the extension (everything after the last dot)
+        displayValue = `${lead}...${trail}${ext}`;
+    }
+
+    return displayValue;
+}
